@@ -7,7 +7,7 @@
     "test": "ng test",
     "lint": "ng lint",
     "preversion": "yarn lint && yarn test",
-    "version": "yarn build",
+    "version": "yarn build --prod",
     "postversion": "git push && git push --tags"
   },
   "repository": {
@@ -26,6 +26,7 @@
     "ng6",
     "ng7",
     "ng8",
+    "ng9",
     "loader",
     "progressbar",
     "spinner",
@@ -37,43 +38,43 @@
     "url": "https://github.com/mpalourdio/ng-http-loader/issues"
   },
   "peerDependencies": {
-    "@angular/common": "^8.0.0",
-    "@angular/core": "^8.0.0",
-    "rxjs": "^6.5.0"
+    "@angular/common": "^9.0.0-rc.6",
+    "@angular/core": "^9.0.0-rc.6",
+    "tslib": "^1.10.0",
+    "rxjs": "^6.5.3"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "~0.803.0",
-    "@angular-devkit/build-ng-packagr": "~0.803.0",
-    "@angular/cli": "^8.0.0",
-    "@angular/common": "^8.0.0",
-    "@angular/compiler": "^8.0.0",
-    "@angular/compiler-cli": "^8.0.0",
-    "@angular/core": "^8.0.0",
-    "@angular/language-service": "^8.0.0",
-    "@angular/platform-browser": "^8.0.0",
-    "@angular/platform-browser-dynamic": "^8.0.0",
-    "@types/jasmine": "~3.3.8",
+    "@angular-devkit/build-angular": "~0.900.0-rc.6",
+    "@angular-devkit/build-ng-packagr": "~0.900.0-rc.6",
+    "@angular/cli": "^9.0.0-rc.6",
+    "@angular/common": "^9.0.0-rc.6",
+    "@angular/compiler": "^9.0.0-rc.6",
+    "@angular/compiler-cli": "^9.0.0-rc.6",
+    "@angular/core": "^9.0.0-rc.6",
+    "@angular/language-service": "^9.0.0-rc.6",
+    "@angular/platform-browser": "^9.0.0-rc.6",
+    "@angular/platform-browser-dynamic": "^9.0.0-rc.6",
+    "@types/jasmine": "~3.5.0",
     "@types/jasminewd2": "~2.0.3",
-    "@types/node": "~8.9.4",
-    "codelyzer": "^5.0.0",
+    "@types/node": "^12.11.1",
+    "codelyzer": "^5.1.2",
     "coveralls": "^3.0.1",
-    "jasmine-core": "~3.4.0",
+    "jasmine-core": "~3.5.0",
     "jasmine-spec-reporter": "~4.2.1",
-    "karma": "~4.1.0",
-    "karma-chrome-launcher": "~2.2.0",
-    "karma-coverage-istanbul-reporter": "~2.0.1",
+    "karma": "~4.3.0",
+    "karma-chrome-launcher": "~3.1.0",
+    "karma-coverage-istanbul-reporter": "~2.1.0",
     "karma-firefox-launcher": "^1.1.0",
     "karma-jasmine": "~2.0.1",
-    "karma-jasmine-html-reporter": "^1.4.0",
-    "ng-packagr": "^5.3.0",
-    "rxjs": "^6.5.0",
+    "karma-jasmine-html-reporter": "^1.4.2",
+    "ng-packagr": "^9.0.0-rc.3",
+    "rxjs": "^6.5.3",
     "rxjs-tslint-rules": "^4.13.0",
-    "ts-node": "~7.0.0",
-    "tsickle": "^0.37.0",
+    "ts-node": "~8.3.0",
     "tslib": "^1.10.0",
-    "tslint": "~5.15.0",
-    "typescript": "~3.5.3",
-    "zone.js": "~0.9.1"
+    "tslint": "~5.18.0",
+    "typescript": "~3.6.4",
+    "zone.js": "~0.10.2"
   },
   "engines": {
     "node": ">= 8.9.0",
